# public/swagger.yaml
openapi: 3.0.0
info:
  title: MagicGames Portal Integration API
  description: |
    This API documentation is for game developers integrating their games into the MagicGames portal.
    
    The portal loads games inside an iframe and passes the `sessionId` as a query parameter.
  version: 1.0.0

servers:
  - url: /api

paths:
  /game/init:
    get:
      summary: Initialize Game Session
      description: Fetches game session details and player profile information.
      parameters:
        - name: sessionId
          in: query
          required: true
          description: The unique ID of the game session.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session details retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionInitResponse'

  /game/match/complete:
    post:
      summary: Submit Match Score
      description: Submits the final score at the end of the session. Must be signed with HMAC-SHA256 signature header.
      parameters:
        - name: X-Portal-Signature
          in: header
          required: true
          description: Hex-encoded HMAC-SHA256 signature of the raw request body signed with the game's webhook_secret.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sessionId
                - score
              properties:
                sessionId:
                  type: string
                  format: uuid
                score:
                  type: integer
      responses:
        '200':
          description: Score submitted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreSubmissionResponse'

components:
  schemas:
    SessionInitResponse:
      type: object
      properties:
        success:
          type: boolean
        session:
          type: object
          properties:
            id:
              type: string
              format: uuid
            mode:
              type: string
              enum: [practice, tournament]
            status:
              type: string
            game_title:
              type: string
            game_slug:
              type: string
        player:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            username:
              type: string
            avatar:
              type: string

    ScoreSubmissionResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        is_personal_best:
          type: boolean
        best_score:
          type: integer